r1488/sofa-dev : Remove too broad 'using namespace' directives in header files. Should fix compilation with Visual C++ 2003...
 //partial sync (55/59 changes)//

+++ modules/sofa/component/topology/RegularGridTopology.cpp
@@ -54,8 +54,8 @@ void RegularGridTopology::parse(core::objectmodel::BaseObjectDescription* arg)
         arg->getAttribute("ymax") != NULL &&
         arg->getAttribute("zmax") != NULL )
     {
-        min.setValue(Vec3f(atof(xmin),atof(ymin),atof(zmin)));
-        max.setValue(Vec3f(atof(xmax),atof(ymax),atof(zmax)));
+        min.setValue(Vec3f((float)atof(xmin),(float)atof(ymin),(float)atof(zmin)));
+        max.setValue(Vec3f((float)atof(xmax),(float)atof(ymax),(float)atof(zmax)));
     }
     this->setPos(min.getValue()[0],max.getValue()[0],min.getValue()[1],max.getValue()[1],min.getValue()[2],max.getValue()[2]);
 }
